More core cpp

+++ src/core/cpp/range.cpp
@@ -3,7 +3,7 @@ shared_ptr<Box>  core_range(shared_ptr<Box> a, shared_ptr<Box> b, shared_ptr<Box
   result->leaf = false;
   result->type = 2;
   int i = 0;
-  for (int n = a->double_leaf; n <= b->double_leaf; n += c->double_leaf) {
+  for (double n = a->double_leaf; n <= b->double_leaf; n += c->double_leaf) {
     result->indexes[i] = make_shared<Box>(n);
     i++;
   }
